// src/common/enums/error-code.enum.ts
export enum ErrorCodeEnum {
  // Standard errors
  NO_CONTENT = 204,
  MOVED_PERMANENTLY = 301,
  FOUND = 302,
  NOT_MODIFIED = 304,
  BAD_REQUEST_EXCEPTION = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  METHOD_NOT_ALLOWED = 405,
  TOO_MANY_REQUESTS = 429,
  SERVER_SIDE_ERROR = 500,
  NOT_IMPLEMENTED = 501,
  BAD_GATEWAY = 502,
  SERVICE_UNAVAILABLE = 503,
  GATEWAY_TIMEOUT = 504,

  // Uploader errors
  // NOTE ERROR CODE FOR THE UPLOADER SHOULD BE LESS THAN 1000
  INVALID_FILE_UPLOAD = 800,
  FILE_TOO_LARGE = 801,
  TOO_MANY_FILES = 802,
  TOO_MANY_FIELDS = 803,
  INVALID_FILE_TYPE = 804,
  INVALID_MODEL_USE_CASE = 805,
  MISSING_CONTENT_TYPE = 806,
  NO_FILE_PROVIDED = 807,
  FILE_DOES_NOT_EXIST = 808,

  //Auth from 1000 to 1299
  EMAIL_ALREADY_EXIST = 1000,
  VALID_VERIFICATION_CODE_EXIST = 1001,
  USER_DOES_NOT_EXIST = 1002,
  USER_EMAIL_ALREADY_VERIFIED = 1003,
  USER_NOT_VERIFIED = 1004,
  INVALID_VERIFICATION_CODE = 1005,
  EXPIRED_VERIFICATION_CODE = 1006,
  WRONG_EMAIL_OR_PASSWORD = 1007,
  VERIFIED_PHONE_NUMBER_EXIST = 1008,
  UNDEFINED_EMAIL_AND_PHONE_NUMBER = 1009,
  USER_EMAIL_DOES_NOT_EXIST = 1010,
  USER_PHONE_DOES_NOT_EXIST = 1011,
  USER_PHONE_NUMBER_ALREADY_VERIFIED = 1012,
  USER_DOES_NOT_HAVE_VERIFIED_EMAIL = 1013,
  USER_DOES_NOT_HAVE_VERIFIED_PHONE_NUMBER = 1014,
  VERIFIED_EMAIL_EXIST = 1015,
  NOT_PROVIDED_EMAIL = 1016,
  SOCIAL_ID_ALREADY_EXIST = 1017,
  EMAIL_PROVIDED_BOTH_MANUALLY_AND_BY_SOCIAL_PROVIDE = 1018,
  INVALID_SOCIAL_AUTH_TOKEN = 1019,
  INCORRECT_PASSWORD = 1020,
  USER_ALREADY_CONNECTED_TO_THIS_SOCIAL_PROVIDER = 1021,
  ADMIN_GROUP_WITH_THIS_NAME_ALREADY_EXIST = 1022,
  INVALID_PERMISSIONS_CODES_LIST = 1023,
  ADMIN_GROUP_DOES_NOT_EXIST = 1024,
  NOT_PROVIDED_NOTIFICATION_TOKEN = 1025,
  EXPIRED_ACCESS_TOKEN = 498,
  PASSWORD_RESET_REQUIRED = 1027,
  PROFILE_COMPLETION_REQUIRED = 1028,
  EMAIL_MISMATCH = 1029,

  //Notifications form 1300 to 1349
  NOTIFICATION_ALREADY_MARKED_AS_SEEN = 1300,
  NOTIFICATION_RECEIVER_DOES_NOT_EXIST = 1301,

  //Region form 1350 to 1399
  COUNTRY_ALREADY_EXIST = 1350,
  INVALID_COUNTRY_CODE = 1351,
  COUNTY_DOES_NOT_EXIST = 1352,

  //payment from 1400 to 1499
  PAYMENT_DOES_NOT_EXIST = 1400,
  CITY_DOES_NOT_EXIST = 1401,

  //other
  INVALID_USER_IDS = 2000,
  FAQ_DOES_NOT_EXIST = 2001,
  APP_CONTACT_DOES_NOT_EXIST = 2002,
  POLICY_WITH_THIS_TYPE_ALREADY_EXIST = 2003,
  POLICY_DOES_NOT_EXIST = 2004,
  PARENT_BLOG_CATEGORY_DOES_NOT_EXIST = 2005,
  NESTED_BLOG_CATEGORY_PARENT = 2006,
  SLUG_ALREADY_EXIST = 2007,
  BLOG_CATEGORY_ALREADY_EXIST = 2008,
  BLOG_CATEGORY_DOES_NOT_EXIST = 2009,
  TAG_DOES_NOT_EXIST = 2010,
}